More refactoring and IDL support
