+++ tool/README.md
@@ -6,9 +6,9 @@ Generates EXE files for some of the Dart samples
 # constants.cpp
 Helper file to confirm widths of structs across various architectures
 
-# generate_com
+# generate.dart
 Used to automatically generate the COM files from C++ fragments. 
 Use with something like the following:
 ```
-dart generate_com\generate.dart generate_com\input ..\lib\src\com
+dart generate\generate.dart generate\com ..\lib\src\com
 ```
